Linux build scripts: Some refinements.
Suppress some pointless compiler warnings for moglcore, moalcore on
recent gcc compiler versions.

Make PsychOpenXRCore and PsychCV build conditional on availability of
OpenXR SDK and Apriltags library. Only avoid PsychOpenXRCore build on
32-Bit ARM, but allow on 64-Bit ARM. Allow PsychCV build on ARM systems
if Apriltags library is available on those systems.

+++ PsychSourceGL/Source/linuxmakeit64.m
@@ -107,7 +107,7 @@ if mode==9
     cd(curdir);
 end
 
-if mode == 10
+if mode == 10 && exist('/usr/include/apriltag', 'dir')
     % Build PsychCV.mexa64:
     mex 'LDFLAGS=$LDFLAGS -Wl,-rpath,\$ORIGIN/PsychPlugins' -L../../Psychtoolbox/PsychBasic/PsychPlugins CFLAGS='$CFLAGS -fPIC -std=gnu99 -fexceptions -pthread' -outdir ../Projects/Linux/build/ -output PsychCV -largeArrayDims -DMEX_DOUBLE_HANDLE -DPTBMODULE_PsychCV -DPSYCHCV_USE_APRILTAGS -I/usr/include/apriltag -ICommon/Base -ILinux/Base -ICommon/PsychCV "Common/Base/*.c" "Linux/Base/*.c" "Common/PsychCV/*.c" -lc -lrt -ldl -lapriltag -lapriltag-utils
     unix(['mv ../Projects/Linux/build/PsychCV.' mexext ' ' PsychtoolboxRoot 'PsychBasic/']);
